refactor(legal): move the legal-link builders out of components/
`services/notifications-settings` renders an upgrade gate too, so importing the
assent renderer from `components/` was a backward layer edge that
`npm run lint:boundaries` rejects. `utils/` is the right home: the module's only
dependency is `escapeHtml`, which already lives there.

No behavior change — same exports, same markup, same tests.

Refs #6976
